Ask confirmation when deleting selector or sitemap (#64)
* Added delete confirmation window. Needed fix bug

* Fixed bag

* Fixed bag of clicking anywhere outside the modal

* Fixed bag of clicking anywhere outside the modal

* Changed the appearance and mechanism of the confirm panel

* Fixed eng messages

* Fixed code style

* Fixed style problems and divided showConfirmPanel function on two (show and wait)

* added spaces

* Fixed styles problems

* Bad fix for backdrop click

* working example

* tried to move i18n keys to view + some renaming

* revert to 'Confirm/Подтвердить' for submit label

* Fixed confirm and cancel buttons

+++ src/scripts/Controller.js
@@ -141,6 +141,7 @@ export default class SitemapController {
 			'SitemapSelectorGraph',
 			'DataPreview',
 			'ItemCard',
+			'ActionConfirm',
 		];
 
 		return Promise.all(
@@ -1196,22 +1197,49 @@ export default class SitemapController {
 			type: 'SelectorText',
 			multiple: false,
 		});
-
 		this._editSelector(selector, sitemap);
 	}
 
+	initConfirmActionPanel(action) {
+		$('#confirm-action-modal').remove(); // remove old panel
+		$('#viewport').after(ich.ActionConfirm(action));
+		Translator.translatePage();
+	}
+
+	showConfirmActionPanel(onConfirm) {
+		const $confirmActionModal = $('#confirm-action-modal');
+		$('#modal-submit').click(() => {
+			$confirmActionModal.modal('hide');
+			onConfirm();
+		});
+		$confirmActionModal.modal('show');
+	}
+
 	async deleteSelector(button) {
-		const sitemap = this.state.currentSitemap;
 		const selector = $(button).closest('tr').data('selector');
-		sitemap.deleteSelector(selector);
-		await this.store.saveSitemap(sitemap);
-		this.showSitemapSelectorList();
+		const sitemap = this.state.currentSitemap;
+		const childCount = sitemap.getDirectChildSelectors(selector.id).length;
+		this.initConfirmActionPanel({ action: 'delete_selector' });
+		$('#modal-selector-id').text(selector.id);
+		if (childCount) {
+			$('#modal-child-count').text(childCount);
+			$('#modal-message').show();
+		}
+		this.showConfirmActionPanel(async () => {
+			sitemap.deleteSelector(selector);
+			await this.store.saveSitemap(sitemap);
+			this.showSitemapSelectorList();
+		});
 	}
 
 	async deleteSitemap(button) {
 		const sitemap = $(button).closest('tr').data('sitemap');
-		await this.store.deleteSitemap(sitemap);
-		await this.showSitemaps();
+		this.initConfirmActionPanel({ action: 'delete_sitemap' });
+		$('#modal-sitemap-id').text(sitemap._id);
+		this.showConfirmActionPanel(async () => {
+			await this.store.deleteSitemap(sitemap);
+			await this.showSitemaps();
+		});
 	}
 
 	initScrapeSitemapConfigValidation() {
